Deregister 'offending' table metrics when table is DROPped
The metrics that are being deregistered (in this PR) cause Scylla to crash when a
table is dropped, but the corresponding table object in memory is not
yet deallocated, and a new table with the same name is created. This
causes double-metrics-registration exception to be thrown. In order to
avoid it, we are deregistering table's metrics as soon as the table is
getting removed from the database. It's representation in memory can
still live, but shouldn't forbid other table with the same name to be
created.