repair_tablets: distribute the primary replica across racks
tablet allocation uses the rack ordinal order
for all tablets to facilitate base/view pairing in the
same rack.  This may lead to a bias when repairing
with the primary-ranges option, since the primary
replica, as currently implemented, will always be
on the same rack.

Instead, calculate a primary rack in each dc using
the tablet id module the number of racks in the dc.
The primary replica would then be the first replica
in that dc/rack.

Since tablet allocation shuffles the node order in
each rack in a uniform distribution, the primary replica
will be uniformly distributed as well.

Fixes scylladb/scylladb#17752

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>