auth: use paged internal queries during migration
Auth v2 migration uses non-paged queries via `execute_internal` API.
This commit changes it to use `query_internal` instead, which uses
paging under the hood.
Note that we lack setting the query timeout now, but this should be
somehow compensated by a paged query, i.e. a limit on the number of
results, which should make the query faster, so theoretically there
should be a smaller chance of hitting the timeout.